+++ src/Dropdown/Dropdown.test.tsx
@@ -23,7 +23,13 @@ const items = [
 
 describe('Dropdown', () => {
     it('renders a list of items', () => {
-        render(<Dropdown items={items} onSelect={jest.fn()} />);
+        render(
+            <Dropdown
+                items={items}
+                onSelect={jest.fn()}
+                selectedItem={items[0]}
+            />
+        );
         expect(screen.getAllByText('Foo').length).toBe(2); // default selected + item
         expect(screen.getByText('Bar')).toBeInTheDocument();
     });
@@ -31,9 +37,26 @@ describe('Dropdown', () => {
     it('calls onSelect when item is clicked', () => {
         const onSelect = jest.fn();
         const item = items[1];
-        render(<Dropdown items={items} onSelect={onSelect} />);
+        render(
+            <Dropdown
+                items={items}
+                onSelect={onSelect}
+                selectedItem={items[0]}
+            />
+        );
         const dropdownItem = screen.getByText(item.label);
         fireEvent.click(dropdownItem);
         expect(onSelect).toHaveBeenCalledWith(item);
+    });
+
+    it('correct item is selected', () => {
+        render(
+            <Dropdown
+                items={items}
+                onSelect={jest.fn()}
+                selectedItem={items[1]}
+            />
+        );
+        expect(screen.getAllByText('Bar').length).toBe(2);
     });
 });